adds module for review rating methods
