+++ app/assets/stylesheets/application.scss
@@ -6,11 +6,12 @@
 
 #map {
   width: 100%;
-  // height: 50%;
   min-height: 250px;
-  // min-width: 100px;
 }
 
+.fa-star, .fa-star-half-o {
+  color: gold;
+}
 
 
 .feed {
