+++ app/models/venue.rb
@@ -1,4 +1,6 @@
 class Venue < ApplicationRecord
+  include ReviewRatingsModule
+
   validates_presence_of :name, :street_address, :city, :state, :zip
   validates_uniqueness_of :name
