+++ app/views/wines/show.html.erb
@@ -7,12 +7,23 @@
   <div class="col s4">
   <div class="profile-overview white-div center-align">
     <h5><%= @wine.name %></h5>
-    <p>Average Rating: <%= @wine.average_review_rating %></p>
     <% if current_user && !current_user.admin? %>
       <%= render partial: "wines/follow/#{current_user.role}_follow" %>
     <% end %>
     <%= render partial: 'wines/wine_details', locals: {wine: @wine} %>
 </div>
+<div class="white-div center-align">
+<h5>Average Rating</h5>
+<%= @wine.average_review_rating %>
+
+<% @wine.average_review_rating.floor.times do %>
+<i class="fa fa-star" aria-hidden="true"></i>
+<% end %>
+<% if @wine.half_star? %>
+  <i class="fa fa-star-half-o" aria-hidden="true"></i>
+<% end %>
+</div>
+
 <div class="white-div center-align">
   <h5>Venues</h5>
   <% @wine.venues.each do |venue| %>
